Get GloVe embedding vector(s) for one or more words

Returns a matrix of 300-dimensional GloVe embeddings for the requested words. Words not present in the selected GloVe matrix are silently dropped (check the row names of the result to see which words were found).

lex_embed(words, model = c("2024", "2014"))

Arguments

words

character vector. One or more words (case-insensitive).

model

character. Which GloVe release to use: "2024" (default) or "2014".

Value

A numeric matrix with one row per found word and 300 columns. Row names are the matched word forms. Returns NULL with a warning if no words are found.

Examples

lex_embed("dog")
#>            V1       V2      V3        V4       V5       V6       V7       V8
#> dog -0.632793 0.139313 0.52401 -0.708676 0.115799 0.399047 0.053947 0.175825
#>            V9      V10       V11      V12     V13       V14       V15       V16
#> dog -0.192239 0.639132 -0.300189 0.828857 -0.2265 -0.844862 -0.252896 -0.435005
#>           V17      V18     V19       V20      V21       V22      V23       V24
#> dog -0.347495 0.434799 0.69983 -0.383584 0.108788 -0.651273 0.218918 -0.246645
#>          V25      V26      V27      V28      V29       V30      V31       V32
#> dog 0.385781 0.508442 0.357227 0.187504 0.276575 -0.317857 0.266694 -0.802001
#>           V33      V34     V35      V36       V37      V38      V39      V40
#> dog -0.278453 0.448305 0.52552 0.136024 -0.814293 0.590886 -0.44105 0.200591
#>          V41      V42       V43       V44      V45       V46      V47       V48
#> dog 0.274412 0.228035 -0.056394 -0.261303 0.210674 -0.257229 -0.15868 -0.043062
#>          V49       V50      V51       V52       V53       V54      V55
#> dog 0.399493 -0.322868 -0.04283 -0.174126 -0.159271 -0.124697 0.080069
#>           V56      V57       V58       V59       V60      V61      V62      V63
#> dog -0.425084 0.256195 -0.018682 -0.003924 -0.022696 0.779844 0.211217 0.000312
#>          V64       V65      V66      V67      V68      V69      V70       V71
#> dog 0.116732 -0.440387 0.020813 -0.08503 0.104897 0.081148 0.223924 -0.573644
#>           V72      V73      V74       V75      V76      V77      V78      V79
#> dog -0.630873 0.407494 0.992513 -0.369787 0.053645 -0.19998 -0.44484 0.068976
#>           V80       V81      V82      V83      V84     V85      V86       V87
#> dog -0.265479 -0.127062 -0.45808 0.318968 0.017424 0.34253 -0.01423 -0.318621
#>          V88       V89      V90       V91      V92      V93      V94       V95
#> dog 0.090739 -0.201105 0.225303 -0.250904 0.028247 0.456043 0.240813 -0.306322
#>          V96      V97       V98       V99      V100      V101     V102
#> dog 0.445561 0.254557 -0.165304 -0.342361 -0.116456 -0.504503 0.483369
#>          V103     V104      V105     V106     V107      V108     V109      V110
#> dog -0.164905 0.046906 -0.654576 0.811403 0.145529 -0.103123 0.271814 -0.586611
#>         V111      V112     V113     V114     V115      V116     V117      V118
#> dog 0.767061 -0.073861 0.595761 0.544009 0.431094 -0.125297 0.213704 -0.343989
#>          V119     V120      V121     V122     V123     V124     V125      V126
#> dog -0.060493 -0.30382 -0.513006 0.250202 0.042042 0.185738 0.242672 -0.130144
#>          V127      V128     V129     V130      V131     V132     V133     V134
#> dog -0.802654 -0.214731 0.155586 0.410134 -0.363131 -0.24292 0.064457 0.041361
#>          V135      V136     V137     V138      V139     V140     V141      V142
#> dog -0.022771 -0.212835 0.018546 0.150556 -0.178016 0.114587 0.298761 -0.120313
#>        V143     V144     V145    V146     V147     V148      V149      V150
#> dog 0.06665 0.259502 -0.32353 0.82176 0.171144 0.452266 -0.557545 -0.260757
#>         V151      V152      V153     V154     V155      V156      V157     V158
#> dog 0.093023 -0.017372 -0.170704 0.256071 0.557992 -0.492528 -0.033322 0.429222
#>         V159      V160      V161     V162     V163     V164     V165      V166
#> dog 0.000781 -0.289799 -0.411656 0.182419 0.763904 0.421284 0.036613 -0.808356
#>         V167      V168      V169      V170      V171     V172     V173     V174
#> dog -0.17069 -0.040622 -0.073839 -0.647959 -0.372304 0.294447 0.389324 0.121513
#>          V175      V176     V177     V178     V179     V180      V181      V182
#> dog -0.569818 -0.098641 0.044627 0.272922 0.015393 0.775685 -0.595561 -0.507909
#>          V183      V184     V185      V186     V187     V188     V189      V190
#> dog -0.221361 -0.157566 -0.04246 -0.253368 0.204279 0.227353 0.229118 -0.300338
#>          V191     V192      V193      V194      V195      V196      V197
#> dog -0.242791 0.162301 -0.382859 -0.147379 -0.780128 -0.037881 -0.263509
#>          V198      V199      V200     V201     V202     V203      V204     V205
#> dog -0.346106 -0.275564 -0.165925 0.190973 0.316678 0.447293 -0.413624 0.075849
#>        V206     V207      V208     V209      V210      V211     V212     V213
#> dog 0.52885 0.388575 -0.640462 0.108037 -0.127393 -0.004951 0.348406 0.284567
#>         V214      V215     V216      V217      V218     V219      V220   V221
#> dog 0.229692 -0.516893 0.333391 -0.380244 -0.199251 0.199706 -0.057324 0.2454
#>          V222      V223      V224     V225      V226     V227      V228
#> dog -0.100674 -0.494216 -0.512094 0.212758 -0.299504 0.169103 -0.476192
#>          V229      V230     V231    V232     V233      V234      V235     V236
#> dog -0.849205 -0.320856 0.273106 0.17429 0.147789 -0.503905 -3.248592 0.104263
#>          V237    V238      V239     V240     V241      V242    V243      V244
#> dog -0.404633 -0.1696 -0.398436 -0.21562 -0.38112 -0.223039 0.15641 -0.168797
#>         V245     V246    V247      V248     V249     V250     V251     V252
#> dog 0.413561 0.393159 0.08869 -0.019359 0.046118 0.116216 -0.11307 0.490183
#>         V253     V254      V255      V256      V257     V258    V259      V260
#> dog 0.024316 0.159194 -0.018992 -0.779736 -0.225294 0.520406 0.04191 -0.112884
#>          V261      V262      V263     V264      V265      V266      V267
#> dog -0.222593 -0.080846 -0.528283 0.137577 -0.018016 -0.026043 -0.309413
#>         V268      V269     V270      V271     V272      V273    V274     V275
#> dog 0.567117 -0.301581 0.287075 -0.091984 0.230266 -0.356407 0.14591 0.418417
#>         V276     V277      V278     V279     V280      V281   V282     V283
#> dog 0.417852 0.360508 -0.060319 0.193336 0.360476 -0.125722 0.4036 0.500161
#>         V284     V285     V286     V287      V288      V289      V290      V291
#> dog -0.27558 0.511126 0.792192 0.267558 -0.387105 -0.347327 -0.390517 -0.143754
#>         V292     V293     V294      V295     V296     V297    V298      V299
#> dog 0.050994 0.170031 0.055377 -0.536975 0.301156 0.322822 0.20744 -0.256417
#>          V300
#> dog -0.166513
lex_embed(c("king", "queen", "man", "woman"))
#>              V1        V2        V3        V4        V5        V6        V7
#> king  -0.069590 -0.370844 -0.210603 -0.881204  0.061031 -0.262035  0.335459
#> queen -0.080426 -0.301578 -0.179378 -0.806072 -0.657794  0.066208  0.214696
#> man   -0.023865 -0.021753 -0.400016  0.267803  0.178442  0.280223 -0.271002
#> woman -0.138734 -0.074131 -0.098307  0.097292 -0.509030  0.025602  0.224397
#>              V8        V9       V10       V11       V12       V13       V14
#> king  -0.226207 -0.156444  0.255534 -0.056904  0.419359  0.023010  0.451314
#> queen -0.338222  0.112106 -0.053744  0.004207  0.069372  0.422947 -0.189757
#> man    0.096707 -0.266119 -0.181009  0.031906 -0.267596 -0.027417  0.374723
#> woman -0.130794 -0.461744  0.025317 -0.062269  0.285774  0.132457  0.106556
#>             V15       V16       V17      V18       V19       V20       V21
#> king  -0.274165  0.416569  0.514095 0.181126 -0.082496 -0.188417 -0.302432
#> queen -0.417046  0.143956  0.028349 0.777127  0.321625 -0.556914 -0.382604
#> man    0.157968  0.348063  0.369580 0.655674 -0.170761  0.101542  0.186351
#> woman -0.281565 -0.080240 -0.168437 0.868812  0.386149 -0.522337  0.086400
#>             V22      V23       V24       V25       V26       V27      V28
#> king  -0.391855 0.070216 -0.299944 -0.125507  0.379253  0.494261 0.233729
#> queen -0.100213 0.400655 -0.016217 -0.335439  0.585494 -0.218057 0.431504
#> man    0.204631 0.156606  0.366475  0.815315 -0.347415 -0.198507 0.101148
#> woman -0.050398 0.324980  0.231075  0.278499 -0.118881  0.095408 0.438389
#>             V29       V30       V31       V32       V33       V34       V35
#> king  -0.314000 -0.102371 -0.354171  0.205862  0.210271 -0.011180 -0.336015
#> queen -0.034259 -0.485304  0.193886 -0.449841  0.223316  0.340895 -0.476850
#> man    0.226284  0.111408 -0.095785  0.055667 -0.134687  0.153626  0.728451
#> woman  0.325745  0.167575  0.186311 -0.654472 -0.152407  0.415904  0.790605
#>            V36       V37       V38       V39       V40       V41       V42
#> king  0.593254 -1.287915 -0.080788 -0.343216 -0.193061  0.227662  0.339030
#> queen 0.975399 -1.203145 -0.099921 -0.196094 -0.106312  0.269109  1.057498
#> man   0.335607 -0.264680 -0.506414  0.023134  0.470205  0.254289 -0.614409
#> woman 0.417535 -0.134187  0.076885  0.129803 -0.045723 -0.019276 -0.008196
#>             V43       V44      V45       V46       V47       V48       V49
#> king  -0.317122 -0.023503 0.241311  0.354832 -0.556036 -0.104316 -0.222916
#> queen -0.253288 -0.002541 0.399520  0.153193 -0.117703 -0.319565  0.195620
#> man   -0.454364 -0.036534 0.002954 -0.231284 -0.252064  0.268362 -0.349313
#> woman -0.514138  0.120654 0.030584  0.503267 -0.085131  0.305954 -0.065715
#>             V50       V51      V52       V53      V54       V55       V56
#> king   0.328626 -0.652584 0.333146 -0.013630 0.114183 -0.099766  0.084818
#> queen  0.120552 -0.872754 0.016085 -0.032134 0.263002  0.038498  0.318736
#> man   -0.381873 -0.020691 0.278256 -0.386067 0.347449  0.094928 -0.186965
#> woman  0.046873 -0.132854 0.319194 -0.500213 0.394558  0.122028  0.263810
#>             V57       V58       V59       V60       V61       V62       V63
#> king  -0.159082 -0.255313  0.140435 -0.396929  0.488012  0.307934 -0.111507
#> queen  0.067121  0.151959  0.390100 -0.309044  0.108086  0.297158 -0.045957
#> man   -0.007749 -0.453505 -0.301551  0.131205  0.240440  0.689151  0.003791
#> woman -0.279968  0.620172 -0.297641  0.221404 -0.005518 -0.020763  0.008899
#>            V64       V65      V66       V67       V68       V69      V70
#> king  0.346354 -0.245206 0.381822 -0.070073 -0.716508 -0.392957 0.164132
#> queen 0.281944 -0.407855 0.501655  0.544241  0.207938 -0.143090 0.659951
#> man   0.352047 -0.125936 0.132289  0.042543  0.033565 -0.310548 0.318422
#> woman 0.318507  0.208996 0.053335  0.052539  0.575845  0.114893 0.142876
#>             V71       V72       V73       V74      V75       V76       V77
#> king  -0.035531  0.210894  0.467627  0.686657 0.473003 -0.259763 -0.437819
#> queen  0.039895  0.105172 -0.058097  0.728625 0.231633 -0.325741 -0.635861
#> man    0.009942 -0.398006  0.055158 -0.078667 0.132450 -0.044404  0.210314
#> woman -0.354340 -0.282488 -0.026949  0.245065 0.657233  0.116587 -0.060345
#>             V78       V79       V80       V81       V82       V83       V84
#> king  -0.248033  0.135515  0.038146 -0.234421  0.090648  0.186624 -0.498381
#> queen  0.008963 -0.183538  0.471848 -0.366129 -0.008601 -0.075532 -0.136386
#> man    0.197228 -0.111452 -0.473553  0.073192 -0.160385  0.374435 -0.105897
#> woman  0.006434 -0.404911 -0.815774  0.005016 -0.106117  0.391499 -0.189266
#>            V85       V86       V87       V88       V89       V90       V91
#> king  0.275510 -0.307883  0.221922 -0.180890 -0.130594  0.206882 -0.054883
#> queen 0.422585 -0.200944  0.131314 -0.889150  0.660019  0.337491  0.201022
#> man   0.363850 -0.256943 -0.186151 -0.179086  0.179643 -0.338112 -0.428627
#> woman 0.219343 -0.505187 -0.258479 -0.630398  0.152177 -0.561545 -0.102336
#>             V92       V93      V94       V95       V96       V97       V98
#> king  -0.298964  0.282905 0.380183  1.155640 -0.140531 -0.328554 -0.112391
#> queen -0.373074  0.212342 0.249150  1.000390  0.249824  0.322569 -0.473542
#> man    0.019128  0.510563 0.169195 -0.154292  0.390355 -0.071359 -0.308072
#> woman -0.248252 -0.007261 0.353679 -0.007377  0.443343  0.451506 -0.370853
#>             V99      V100      V101      V102      V103      V104      V105
#> king  -0.432641 -0.181905  0.107615 -0.114631 -0.098659 -0.210332 -0.321028
#> queen  0.058578 -0.056775 -0.119316 -0.061527 -0.030524  0.384921  0.118632
#> man   -0.339551  0.131477  0.456131  0.123406 -0.379988  0.103446  0.006035
#> woman  0.141054 -0.087396  0.436634 -0.438101 -0.133502 -0.093753  0.078790
#>            V106      V107      V108      V109      V110      V111      V112
#> king  -0.873105 -0.076098 -0.268243 -0.104699  0.707693  0.252192 -0.317245
#> queen -0.585086 -0.464609 -0.123224 -0.157348  0.262599 -0.007252 -0.320977
#> man    0.264301 -0.132744 -0.128714 -0.003100 -0.433682  0.195895  0.092681
#> woman  0.506976 -0.692625  0.279561  0.168962 -0.398298 -0.010331  0.800109
#>            V113      V114     V115      V116     V117      V118      V119
#> king   0.334301 -0.306449 0.795797  0.383045 0.377259  0.089376 -0.646629
#> queen -0.010030  0.143876 0.254998 -0.082687 0.353586 -0.378491 -0.887126
#> man    0.567988 -0.193547 0.239183 -0.044244 0.577691  0.033304  0.403035
#> woman  0.559736 -0.205611 0.166994 -0.148610 0.231979 -0.325055  0.338148
#>            V120      V121     V122      V123      V124      V125      V126
#> king  -0.264237  0.069854 0.056683 -0.360254 -0.083936 -0.403487  0.460177
#> queen -0.494503 -0.158863 0.472909 -0.812516  0.360151  0.221156 -0.154404
#> man   -0.011486 -0.250681 0.488002  0.132346 -0.020868  0.026292  0.393427
#> woman  0.166126 -0.332959 0.496836 -0.172363 -0.459909  0.059296  0.466140
#>            V127      V128      V129      V130      V131      V132     V133
#> king  -0.155868  0.642978  0.325797 -0.065293  0.403938  0.373881 0.652628
#> queen -0.109956  0.482638  0.223053 -0.171600 -0.323709  0.215542 0.536660
#> man   -0.642758 -0.246637 -0.073220  0.110843  0.074145  0.079990 0.408149
#> woman -0.298168  0.077429 -0.068558 -0.014481 -0.165893 -0.032888 0.013399
#>            V134      V135      V136      V137      V138      V139      V140
#> king  -0.361319 -0.024115 -0.055430 -0.082667  0.121924 -0.551839  0.147636
#> queen  0.020567 -0.225853  0.100702 -0.208104  0.047031  0.287087  0.197764
#> man   -0.024341  0.675758 -0.228214  0.136677 -0.170628  0.318655 -0.035194
#> woman  0.158185  0.066584 -0.283956  0.743420 -0.348958  0.693166  0.024781
#>            V141      V142      V143      V144      V145      V146      V147
#> king  -0.299508 -0.182830 -0.188320  0.423951  0.243470  0.022163 -1.175083
#> queen -0.555298 -0.044196 -0.245056  0.129033 -0.371009 -0.203807 -0.908573
#> man    0.229076  0.188884  0.027154  0.180826  0.372800  0.213538 -0.291958
#> woman -0.076332  0.290653 -0.134090 -0.407257  0.359464  0.339721 -0.241443
#>            V148      V149      V150      V151     V152      V153      V154
#> king   0.012137 -0.757584 -0.547998 -0.380786 0.637875  0.912072 -0.610393
#> queen -0.558211 -0.545558 -0.627726 -0.303641 1.161188  0.736412 -0.344054
#> man    0.382517 -0.350711 -0.320790 -0.047023 0.116801 -0.245291  0.066199
#> woman  0.213929 -0.352346 -0.411777 -0.169693 0.750788 -0.191813 -0.039775
#>            V155      V156      V157      V158      V159     V160      V161
#> king   0.834348 -0.096988 -0.064778 -0.019221  0.050941 0.022062 -0.604208
#> queen  0.446323  0.069398  0.125096 -0.313714  0.072383 0.175715 -0.570309
#> man   -0.034046 -1.224784 -0.113632 -0.148569 -0.390493 0.325519 -0.128755
#> woman -0.205092 -1.232235  0.143686 -0.042714 -0.036628 0.259416 -0.175790
#>            V162      V163     V164      V165      V166     V167      V168
#> king   0.212074  0.490198 0.163638 -0.211538 -0.380375 0.408273 -0.516723
#> queen -0.063254 -0.209442 0.010419  0.127554  0.181757 0.519037  0.477021
#> man    0.638602  0.638942 0.361282  0.161451 -0.168794 0.152937  0.169556
#> woman  0.648538  0.237928 0.517475  0.018719  0.262653 0.276237  0.651249
#>           V169      V170      V171      V172      V173      V174      V175
#> king  0.044786 -0.024066  0.477167  0.707394  0.035707 -0.040610  0.536469
#> queen 0.320333 -0.192695  0.312259  0.189262 -0.286790 -0.205783  0.275327
#> man   0.322036 -0.276284 -0.164274  0.172457 -0.098706  0.289878 -0.388436
#> woman 0.258979 -0.265463 -0.195760 -0.115764 -0.326303 -0.112946 -0.463518
#>            V176      V177      V178      V179      V180      V181     V182
#> king  -0.315124 -0.272836 -0.059394  0.285520 -0.053771 -0.122262 0.100369
#> queen  0.018110  0.096631  0.170666  0.015824  0.086374  0.321835 0.413262
#> man   -0.529023 -0.308510  0.346645  0.024803 -0.080364  0.010577 0.309584
#> woman -0.172395 -0.226845  0.563163 -0.321181 -0.070842 -0.037026 0.096968
#>            V183      V184     V185      V186      V187      V188      V189
#> king  -0.208105 -0.275223 0.112959 -0.433283 -0.084992 -0.271659  0.481974
#> queen -0.066967 -0.779883 0.111424  0.083284 -0.299216 -0.946086 -0.273077
#> man   -0.436051  0.038040 0.532990 -0.149849 -0.037973 -0.282493  0.225556
#> woman -0.366266 -0.167563 0.175451 -0.104452 -0.023206 -0.642022 -0.226733
#>            V190      V191      V192     V193      V194      V195     V196
#> king   0.420859  0.605329  0.404392 0.114082  0.340643 -1.092833 0.335846
#> queen  0.023466 -0.076574 -0.109194 0.186535 -0.299395 -1.300397 0.511680
#> man   -0.627781 -0.174616  0.114892 0.368809 -0.136564 -0.299962 0.120571
#> woman  0.033287 -0.520776 -0.268020 0.373874 -0.415937 -0.523150 0.203521
#>            V197      V198     V199     V200      V201      V202      V203
#> king  -0.370478  0.501533 0.266268 0.702625 -0.203094  0.020567 -0.075701
#> queen -0.192828  0.492352 0.490663 0.493099 -0.280771 -0.350138 -0.300679
#> man    0.036715 -0.268424 0.064741 0.211825  0.069223  0.257268  0.354145
#> woman  0.057819 -0.108926 0.403452 0.070410 -0.074774 -0.080741  0.004116
#>            V204      V205      V206      V207      V208      V209      V210
#> king  -0.442644 -0.775954  0.184013  0.314512 -0.273776  0.480044 -0.233109
#> queen -0.425087 -0.381734 -0.374393  0.399186 -1.219184  0.564743 -0.387471
#> man   -0.340732 -0.042921 -0.221244 -0.342418 -0.789527 -0.152775  0.051212
#> woman -0.062781 -0.138810 -0.716502 -0.043439 -0.828053  0.130752  0.043387
#>            V211      V212     V213      V214      V215      V216      V217
#> king  -0.341839 -0.473659 0.406493 -0.528930 -0.234033 -0.208162  0.129000
#> queen -0.380752 -0.285122 0.341574 -0.787645  0.016312  0.080046  0.041060
#> man   -0.031015  0.184178 0.234938 -0.120081 -0.184689 -0.062224  0.040692
#> woman -0.353240  0.014663 0.015996  0.225265  0.001291 -0.138829 -0.234813
#>            V218      V219     V220      V221      V222      V223      V224
#> king   0.151133  0.152981 0.319918 -0.031834 -0.324714  0.422633 -0.165108
#> queen -0.107123  0.191458 0.084331  0.114127 -0.522164  0.346718 -0.032348
#> man   -0.187036 -0.290380 0.222637 -0.092497 -0.355396 -0.018711 -0.063439
#> woman  0.020304 -0.254376 0.369618 -0.038787 -0.380606 -0.229536  0.312527
#>            V225      V226     V227      V228     V229      V230      V231
#> king  -0.108770  0.236132 0.811151 -0.553059 0.358255 -0.275389 -0.190106
#> queen -0.195024  0.478738 0.938706 -0.112861 0.648865 -0.141766  0.261656
#> man   -0.325228 -0.104808 0.027383 -0.389596 0.036077  0.171959  0.068963
#> woman -0.261528  0.036966 0.100586 -0.245097 0.342377  0.143579  0.054842
#>           V232      V233      V234      V235      V236      V237      V238
#> king  0.207547  0.636894  0.280680 -4.383976 -0.340963 -0.585035  0.177133
#> queen 0.160944  0.161190  0.054576 -3.456575 -0.026225 -0.734413  0.379256
#> man   0.219752 -0.121964 -0.086509 -4.886720  0.432579 -0.434341 -0.726255
#> woman 0.052161 -0.222148  0.098844 -4.447075  0.586132 -0.676396  0.046581
#>            V239      V240      V241      V242     V243      V244      V245
#> king   0.208776 -0.141820 -0.582404 -0.033554 0.084862 -0.165648 -0.107847
#> queen  0.144906 -0.064570 -0.511990 -0.245716 0.073081 -0.076459 -0.152487
#> man   -0.788062 -0.056608  0.287355  0.146906 0.087536  0.420288 -0.265049
#> woman -0.376255 -0.236868  0.167734 -0.185591 0.339413 -0.029096  0.108167
#>            V246      V247      V248      V249      V250      V251      V252
#> king  -0.205322 -0.097061 -0.225938 -0.004099 -0.217441 -0.337494  0.444242
#> queen -0.149935 -0.216210 -0.233469  0.342915  0.171093 -0.464961  0.138698
#> man   -0.056960  0.026501 -0.239189  0.361934  0.398201 -0.353803  0.048754
#> woman -0.242889  0.214204 -0.349612  0.300875  0.580485 -0.479855 -0.286674
#>            V253      V254      V255      V256      V257      V258     V259
#> king   0.300936 -0.203473 -0.383595 -0.311066  0.096696 -0.379055 0.545889
#> queen -0.289436 -0.044160  0.120116 -0.496620  0.236080  0.398592 0.654155
#> man    0.157296 -0.255836  0.039682  0.126151 -0.222305  0.151609 0.281230
#> woman  0.144704  0.417269  0.068973 -0.266712  0.276911  0.068665 0.081327
#>            V260      V261     V262      V263      V264     V265      V266
#> king   0.226022  0.467900 0.081080 -0.090006 -0.478093 0.062963 -0.027688
#> queen -0.308904 -0.143779 0.270553 -0.103627  0.235469 0.423881 -0.222158
#> man    0.007578 -0.051105 0.319985  0.041553  0.253110 0.009732  0.149497
#> woman -0.255478 -0.275678 0.397789 -0.443882  0.328001 0.158010  0.076180
#>            V267     V268      V269      V270      V271      V272      V273
#> king  -0.385443 0.320993  0.095016 -0.000342  0.458587 -0.649062  0.286170
#> queen  0.186679 0.152988  0.352909  0.018843  0.234611  0.086304  0.499018
#> man   -0.161558 0.276231 -0.241169  0.083980 -0.037897  0.260736 -0.228360
#> woman -0.000520 0.078701 -0.314041 -0.174583 -0.050682  0.126125 -0.084522
#>            V274      V275      V276      V277      V278      V279      V280
#> king   0.286662  0.186664 -0.574659  0.633578  0.108145 -0.198837 -0.384153
#> queen -0.195233  0.309121 -0.679197 -0.035106 -0.466836  0.061099 -0.608390
#> man   -0.097862  0.106219  0.016598  0.301552 -0.022965 -0.003517  0.208737
#> woman -0.348365 -0.194590  0.211971 -0.267925 -0.466126  0.242385 -0.276911
#>            V281      V282      V283     V284      V285      V286      V287
#> king  -0.100449  0.612485 -0.399409 0.326397 -0.381609  0.041447  0.178816
#> queen -0.011588 -0.408914 -0.602211 0.501170 -0.448965 -0.005455 -0.126889
#> man    0.150367  0.463143  0.215857 0.023140  0.212794  0.595830  0.318467
#> woman  0.460486  0.200175  0.592010 0.199037  0.096338  0.444791  0.138261
#>           V288      V289      V290      V291     V292      V293      V294
#> king  0.063388  0.063371 -1.059506 -0.477163 0.191453  0.156133 -0.204406
#> queen 0.078235  0.378402 -0.520229 -0.570924 0.206520  0.708703 -0.181733
#> man   0.072635 -0.372874 -0.515728 -0.324691 0.128748 -0.115962 -0.006867
#> woman 0.588482 -0.429657 -0.683906  0.039036 0.386379  0.079321  0.029234
#>           V295      V296      V297      V298      V299      V300
#> king  0.534611  0.109829 -0.449090  0.193206 -0.273817 -0.004340
#> queen 0.530440  0.518172 -0.498779 -0.021832 -0.412041 -0.182365
#> man   0.158921  0.049313 -0.551194 -0.061119 -0.090400  0.040188
#> woman 0.298704 -0.026746 -0.408879 -0.010369  0.115487  0.340826
lex_embed("dog", model = "2014")
#>           V1      V2       V3      V4        V5       V6      V7      V8
#> dog -0.11043 0.81217 0.073668 0.19023 -0.052888 0.061468 0.16076 0.41302
#>           V9      V10     V11      V12      V13     V14     V15     V16     V17
#> dog -0.30199 -0.90827 0.27504 -0.03189 -0.28842 0.23447 0.47679 0.50124 0.29371
#>         V18      V19      V20     V21     V22      V23     V24      V25
#> dog 0.27029 0.054745 0.098038 0.57116 0.36755 0.040734 0.34347 -0.18256
#>          V26      V27      V28     V29     V30      V31      V32      V33
#> dog -0.28935 0.023826 -0.19401 0.24444 0.13407 -0.16494 -0.26983 -0.26234
#>          V34      V35     V36       V37       V38     V39     V40      V41
#> dog -0.21779 -0.87528 0.73822 -0.087931 -0.010876 -0.2654 0.34668 -0.55814
#>         V42     V43      V44     V45    V46     V47       V48      V49      V50
#> dog 0.17591 0.16926 -0.15725 -0.5043 -0.201 0.66701 -0.032518 0.045012 0.065675
#>          V51      V52     V53     V54     V55      V56      V57       V58
#> dog -0.16061 -0.73363 0.24642 0.34325 0.21899 0.048646 -0.59987 -0.058153
#>           V59      V60 V61     V62     V63        V64     V65      V66
#> dog -0.051694 -0.57846 0.3 0.35078 0.46646 -0.0075309 0.10455 -0.51016
#>           V67      V68      V69      V70       V71     V72      V73      V74
#> dog -0.055987 -0.10295 -0.26476 -0.04123 -0.028371 0.51979 -0.34849 -0.47217
#>          V75      V76     V77     V78     V79      V80     V81     V82     V83
#> dog -0.37229 -0.03279 0.13989 0.35716 0.19305 -0.21986 0.24136 0.40976 0.37516
#>         V84       V85      V86      V87     V88      V89      V90      V91
#> dog 0.14255 -0.034143 -0.72653 -0.10832 0.68616 -0.26335 -0.42345 -0.24253
#>         V92     V93      V94      V95     V96     V97     V98     V99     V100
#> dog 0.15778 0.14258 -0.32749 -0.34699 0.16148 0.19603 0.41639 -0.2337 0.075816
#>        V101      V102      V103     V104     V105    V106     V107    V108
#> dog 0.15899 0.0016623 -0.048301 -0.10611 -0.19326 0.14494 0.015406 0.10629
#>          V109   V110    V111    V112    V113     V114     V115    V116     V117
#> dog -0.036699 0.6323 0.12986 0.49902 -1.1323 -0.12636 0.064718 0.12374 -0.49712
#>          V118    V119     V120     V121    V122      V123     V124      V125
#> dog -0.014836 0.10488 -0.49818 -0.28856 0.38949 -0.031828 -0.28625 -0.098758
#>         V126     V127    V128    V129    V130    V131     V132     V133    V134
#> dog -0.07699 -0.24234 0.75793 0.34835 -0.7103 0.45318 -0.34418 -0.19459 0.61478
#>         V135     V136    V137    V138    V139     V140   V141    V142     V143
#> dog -0.02901 -0.27864 0.38556 0.10072 0.12895 0.017992 0.3367 0.20698 -0.38049
#>           V144   V145      V146     V147    V148      V149    V150     V151
#> dog -0.0066661 0.1154 -0.085268 -0.14608 0.44514 -0.093674 0.23639 -0.11447
#>       V152      V153     V154   V155      V156      V157    V158     V159
#> dog 1.0948 -0.057823 -0.16295 0.5588 -0.018988 -0.071374 0.21319 0.061277
#>        V160    V161    V162    V163    V164     V165    V166  V167     V168
#> dog 0.72759 0.62747 -0.1928 0.13057 0.17426 -0.10229 0.15232 0.525 -0.21919
#>         V169     V170    V171    V172     V173    V174     V175     V176
#> dog -0.27185 -0.54186 0.31752 0.16375 -0.29039 0.17074 -0.31814 -0.96421
#>        V177     V178    V179     V180    V181     V182     V183     V184
#> dog -0.1161 -0.29951 0.18686 -0.45986 0.41633 -0.17583 -0.34583 -0.27244
#>         V185     V186    V187     V188    V189     V190     V191     V192
#> dog -0.50216 0.012852 0.59838 -0.11237 0.24697 -0.49048 -0.44188 -0.16255
#>         V193     V194     V195     V196     V197     V198      V199     V200
#> dog -0.73313 -0.37677 -0.68925 0.061174 -0.42101 -0.13153 -0.008359 -0.01836
#>       V201     V202    V203      V204     V205    V206    V207    V208     V209
#> dog 1.3686 0.046169 0.94622 -0.015126 -0.12477 0.48754 0.22384 -0.2182 -0.23389
#>        V210     V211     V212     V213     V214     V215    V216     V217
#> dog 0.15207 -0.28718 -0.63908 -0.22383 -0.18014 -0.33548 0.53587 -0.29367
#>        V218     V219       V220     V221    V222    V223     V224      V225
#> dog 0.10866 0.063411 -0.0093424 -0.15886 0.22602 0.11925 -0.41442 -0.078062
#>          V226    V227     V228     V229    V230    V231     V232     V233
#> dog -0.097857 0.27938 -0.18348 -0.34584 0.18489 0.17402 -0.52198 -0.43306
#>        V234    V235    V236    V237     V238     V239    V240    V241      V242
#> dog 0.16256 0.14032 0.35124 -0.1828 -0.35984 -0.13009 0.16304 0.31734 0.0037716
#>          V243     V244     V245     V246     V247     V248     V249    V250
#> dog -0.045498 -0.42066 -0.44419 -0.68985 -0.49359 0.070281 -0.14377 0.62508
#>          V251   V252      V253    V254   V255    V256    V257     V258     V259
#> dog -0.056311 0.1885 -0.056785 0.14052 1.1973 0.71894 0.54332 -0.12461 -0.11978
#>        V260     V261     V262     V263      V264     V265    V266    V267
#> dog 0.30163 -0.16273 -0.04674 -0.25249 -0.030659 -0.32271 0.32361 0.33244
#>          V268     V269      V270     V271     V272     V273     V274     V275
#> dog -0.027819 -0.33367 -0.023444 -0.50394 -0.20587 -0.13013 -0.35884 0.045384
#>         V276    V277    V278     V279    V280     V281    V282    V283    V284
#> dog -0.11863 -1.7257 0.39441 -0.53179 0.58209 -0.65771 0.36849 0.23518 0.10802
#>         V285    V286    V287     V288    V289     V290     V291    V292
#> dog -0.83159 0.61486 0.25547 -0.45289 0.51446 -0.17911 -0.12389 0.18688
#>         V293     V294     V295     V296   V297    V298    V299     V300
#> dog -0.41102 -0.70877 -0.37501 -0.66152 0.6773 0.33936 0.57994 0.068149